requirements.txt
discord.py
SQLAlchemy
aiosqlite
